consolidate test setup by moving helper functions to conftest.py
